fix(report article): fix admin permission
