+++ src/middleware/adminPermissions.js
@@ -1,3 +1,14 @@
+const checkAdmin = (req, res, next) => {
+  const userRole = req.userData.role;
+
+  if (userRole !== 'admin' && userRole !== 'super-admin') {
+    return res.status(403).json({
+      error: 'Unathorized access'
+    });
+  }
+  next();
+};
+
 const adminPermission = (req, res, next) => {
   const userRole = req.userData.role;
 
@@ -9,4 +20,4 @@ const adminPermission = (req, res, next) => {
   next();
 };
 
-export default adminPermission;
+export { adminPermission, checkAdmin };
