+++ src/routes/routes.js
@@ -11,7 +11,7 @@ import facebookRequest from '../middleware/facebook';
 import UserFollow from '../controllers/followController';
 import Profile from '../controllers/profileController';
 import canEditProfile from '../middleware/editProfile';
-import adminPermissions from '../middleware/adminPermissions';
+import { adminPermission, checkAdmin } from '../middleware/adminPermissions';
 import Article from '../controllers/articleController';
 import articleRate from '../controllers/ratingController';
 import ArticleMiddleware from '../middleware/articleMiddleware';
@@ -20,14 +20,14 @@ import findOwner from '../middleware/findOwner';
 import Bookmark from '../controllers/bookmarkController';
 import findUser from '../middleware/findUser';
 import Report from '../controllers/reportController';
-import admin from '../middleware/checkAdmin';
+
 
 const router = express.Router();
 
 const connectMulti = connectmultiparty();
 
 router.post('/api/v1/users/signup', Validation.signupValidation, UserAuth.signup);
-router.post('/api/v1/users', verifyToken, adminPermissions, Validation.signupValidation, UserAuth.signup);
+router.post('/api/v1/users', verifyToken, adminPermission, Validation.signupValidation, UserAuth.signup);
 router.get('/api/v1/profiles/:username', Profile.user);
 router.get('/api/v1/profiles', verifyToken, Profile.fetchProfiles);
 router.put('/api/v1/profiles/:username', [verifyToken, connectMulti, canEditProfile, Validation.profileValidation, Validation.imageValidation], Profile.editProfile);
@@ -49,7 +49,7 @@ router.get('/api/v1/auth/twitter/callback', passport.authenticate('twitter'), Us
 router.post('/api/v1/users/logout', [verifyToken], UserAuth.logout);
 router.post('/api/v1/users/login', Validation.loginValidation, UserAuth.login);
 router.delete('/api/v1/users/:username', verifyToken, UserAuth.deleteUser);
-router.patch('/api/v1/users/:username', verifyToken, adminPermissions, Validation.updateRoleValidation, UserAuth.updateRole);
+router.patch('/api/v1/users/:username', verifyToken, adminPermission, Validation.updateRoleValidation, UserAuth.updateRole);
 router.get('/api/v1/articles/:slug', Article.getSingleArticle);
 router.post('/api/v1/articles', [verifyToken, connectMulti, Validation.createArticleValidation, ContentType, Validation.imageValidation], Article.createArticle);
 router.get('/api/v1/articles', Article.getAllArticles);
@@ -75,9 +75,9 @@ router.get('/api/v1/bookmarks', [verifyToken, findUser], Bookmark.getBookmarks);
 router.delete('/api/v1/bookmarks/:slug', [verifyToken, findUser], Bookmark.deleteBookmark);
 
 router.post('/api/v1/articles/:articleSlug/reports', [verifyToken, Validation.reportValidation], Report.createReport);
-router.get('/api/v1/articles/reports/all', [verifyToken, admin], Report.getAllReports);
-router.get('/api/v1/articles/:articleSlug/reports', [verifyToken, admin], Report.getArticleReports);
+router.get('/api/v1/articles/reports/all', [verifyToken, checkAdmin], Report.getAllReports);
+router.get('/api/v1/articles/:articleSlug/reports', [verifyToken, checkAdmin], Report.getArticleReports);
 router.delete('/api/v1/articles/:articleSlug/reports/:reportId', verifyToken, Validation.reportParamsValidation, Report.deleteReport);
-router.get('/api/v1/articles/:articleSlug/reports/:reportId', [verifyToken, admin, Validation.reportParamsValidation], Report.getSingleReport);
+router.get('/api/v1/articles/:articleSlug/reports/:reportId', [verifyToken, checkAdmin, Validation.reportParamsValidation], Report.getSingleReport);
 
 export default router;
